Feature - Player Taunt (#11)
* Created taunt input bindings

Created the Taunt and CancelTaunt input bindings that so far print a taunt to the screen until the taunt button is released

* Implemented basic taunt mechanic

When the player holds the controller face left button or the F keyboard key, the character will automatically chain taunts one after another until the same button is released or the dash ability is triggered.

* Implemented BlueprintImplementable events

Left behind comments in the Blueprint graph to describe the use cases of the PlayTauntEffects and CancelTauntEffects BlueprintImplementable events

* Added extra prints for better taunt responsivity

Based on the randomized taunt type, the taunt implementation will now print a different taunt to the screen for easier preliminary playtesting. This should be removed when the animation and SFX implementations come in

* Fixed disabled taunt on start bug

* Disabled movement during a taunt

When executing a taunt chain, the player's movement will be frozen. The only way to exit the taunt chain while holding the taunt button is to trigger a dash command

* Organized and cleaned code for better readability

* Fixed invalid class keywords