Feature - Enemy Patrol AI (#12)
* Created taunt input bindings

Created the Taunt and CancelTaunt input bindings that so far print a taunt to the screen until the taunt button is released

* Implemented basic taunt mechanic

When the player holds the controller face left button or the F keyboard key, the character will automatically chain taunts one after another until the same button is released or the dash ability is triggered.

* Implemented BlueprintImplementable events

Left behind comments in the Blueprint graph to describe the use cases of the PlayTauntEffects and CancelTauntEffects BlueprintImplementable events

* Added extra prints for better taunt responsivity

Based on the randomized taunt type, the taunt implementation will now print a different taunt to the screen for easier preliminary playtesting. This should be removed when the animation and SFX implementations come in

* Fixed disabled taunt on start bug

* Disabled movement during a taunt

When executing a taunt chain, the player's movement will be frozen. The only way to exit the taunt chain while holding the taunt button is to trigger a dash command

* Organized and cleaned code for better readability

* Fixed invalid class keywords

* Updated the DemoLevel map to include obstacles

Updated the DemoLevel map with obstacles to later test the enemy navigation when developing the AI

* Set up enemy navmesh to navigate DemoLevel

Set the navmesh bounds volume for the DemoLevel to support enemy navigation around obstacles and elevated platforms

* Set up basic charge enemy AI BT

Implemented a basic behaviour tree structure for the charge enemy and imported the requisite modules for the system to work correctly

* Created Enemy module to encapsulate enemy logic

Set up the Enemy module to encapsulate all logic pertaining to the enemies such as BT AI, enemy stats logic, etc. and shield strict enemy concerns from the core module (Player logic, GameMode, GameState/World)

* Implemented Enemy AIController

Implemented the Enemy AIController to possess the Enemy Character class and set and run the associated BehaviorTree

* Implemented random position logic BT task

Implemented the DetermineRandomPosition BT Task to have the Enemy Character execute the BehaviorTree asset at runtime, continuously choosing a random location, moving to it, and waiting. Also updated the DemoLevel map to include the enemy character BP assets for testing.

* Created test animation assets for the enemies

* Organized and documented Enemy module

* Organized scripts in the Enemy module

Created a new hierarchy with subdirectories within the enemy module for better organization and ease of searching